rename resolvers' files
