+++ hima/experiments/temporal_pooling/new/resolvers/dataset.py
@@ -5,7 +5,7 @@
 #  Licensed under the AGPLv3 license. See LICENSE in the project root for license information.
 
 from hima.common.config_utils import extracted_type, resolve_init_params, TConfig
-from hima.experiments.temporal_pooling.new.resolvers.graph_resolver import BlockResolver
+from hima.experiments.temporal_pooling.new.resolvers.graph import BlockResolver
 from hima.experiments.temporal_pooling.new.blocks.graph import Block
