+++ hima/experiments/temporal_pooling/new/resolvers/encoder.py
