+++ hima/experiments/temporal_pooling/new/resolvers/sp.py
@@ -3,9 +3,10 @@
 #  All rights reserved.
 #
 #  Licensed under the AGPLv3 license. See LICENSE in the project root for license information.
+
 from hima.common.config_utils import TConfig, resolve_init_params
 from hima.experiments.temporal_pooling.new.blocks.graph import Block
-from hima.experiments.temporal_pooling.new.resolvers.graph_resolver import BlockResolver
+from hima.experiments.temporal_pooling.new.resolvers.graph import BlockResolver
 
 
 class SpatialPoolerResolver(BlockResolver):
